Added a valid Coordinate check
Implemented 4 Unit Tests for making sure a coordinate is valid.

- Can't be less than 0
- Can't be greater than the max grid size
- Grid size is one smaller technically, as the grid starts at 0,0

Implemented the test logic into the Input Controller, also added method to get the resized grid. With each individual check, will be able to display an error message according to what the error is, when a false is triggered.

To Do:
- Have event if text is a valid command
- Have the textbox linked up, so UIController will run when enter or button is pressed
- Have strings added to the command box