Input Controller Tests + Implementation
Created tests for making sure a PLACE command is valid and that string can be broken apart for CommandBox (takes in multiple commands).

- Splitting strings by the /n char
- Making sure PLACE commands have coordinates and direction value
- Realised an issue with the Get Coordiantes for PLACE command test, where I was using an unfiltered string, which wouldnt be the case as all texts are filtered by this point).

Implemented the code from the tests. Linked up all the logic with the UI objects. Added the Scriptable Object event system and tested to make sure it works.

To Do:
- Implement the remaing events for each Command
- Merge UI branch to Master
- Make Unit test for Robot code
- Tidy up code - will need a SOVector2 for the Grid