Setup Manager Implemented + Game Tested
Implemented the setup manager, which setups up the Robot, Grid and Direction Dictionary. It also checks to make sure a Robot and Grid exist in the scene.

Also tested the scene, there was 2 errors - I had assigned the wrong Event to the Place listener, so it wasn't listening for the raise place event (this has been swapped). I also found that when I was executing the string of multiple commands, I didn't have a check for the Place command, so a PLACE2,3,S wouldnt be found in the dictionary, added a check to see if the string contained PLACE and if it did find the PLACE method in the dictionary instead. Otherwise just find the command in the dictionary like normal.

Found in testing:
- Report text needs to rotate to face the camera